+++ bukkit/src/main/java/pl/bestguilds/bukkit/BukkitBestGuilds.java
@@ -1,9 +1,7 @@
 package pl.bestguilds.bukkit;
 
 import org.bukkit.plugin.java.JavaPlugin;
-import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
-import pl.bestguilds.BestGuilds;
 import pl.bestguilds.BestGuildsPlugin;
 import pl.bestguilds.api.BestGuildsAPI;
 import pl.bestguilds.api.command.CommandInjector;
@@ -11,25 +9,22 @@ import pl.bestguilds.api.command.CommandManager;
 import pl.bestguilds.api.guild.GuildManager;
 import pl.bestguilds.api.user.UserManager;
 import pl.bestguilds.bukkit.command.BukkitCommandInjector;
-import pl.bestguilds.bukkit.util.Updater;
+import pl.bestguilds.Updater;
 
 public final class BukkitBestGuilds extends JavaPlugin implements BestGuildsAPI {
 
   private BestGuildsPlugin plugin;
-  private Updater updater;
+
+  @Override
+  public void onLoad() {
+    this.plugin = new BestGuildsPlugin();
+    Updater.checkVersion(this.getLogger(), this.getDescription().getVersion());
+  }
 
   @Override
   public void onEnable() {
-    this.plugin = new BestGuildsPlugin();
     CommandInjector commandInjector = new BukkitCommandInjector(this);
     this.plugin.registerCommands(commandInjector);
-    this.updater.init(this);
-    try {
-      this.updater.checkForUpdates();
-    } catch (Exception e) {
-      e.printStackTrace();
-    }
-
   }
 
   @Override
@@ -40,18 +35,18 @@ public final class BukkitBestGuilds extends JavaPlugin implements BestGuildsAPI 
   @NotNull
   @Override
   public final UserManager getUserManager() {
-    return this.plugin.getUserManager();
+    return plugin.getUserManager();
   }
 
   @NotNull
   @Override
   public final GuildManager getGuildManager() {
-    return this.plugin.getGuildManager();
+    return plugin.getGuildManager();
   }
 
   @NotNull
   @Override
   public final CommandManager getCommandManager() {
-    return this.plugin.getCommandManager();
+    return plugin.getCommandManager();
   }
 }
